Numerous changes to get to a working state.
GeneticManager: Tweaked logic for finding specific agents from ID. Quite
a bit of refactoring/optimization could be done

EnvironmentProcessor: Added a quick hack for outerloop3. Should get
addressed more formally.

GeneticNNMain: New main file

PlayOurs.sh: Runner for our new main method.

IProjectManager: Few minor changes

Projectmanager: More implementation

+++ src/ch/idsia/agents/reinforcementlearningcs175/IProjectManager.java
@@ -31,7 +31,7 @@ public interface IProjectManager
 
 	public int getLevelSeed();
 
-	public void setLevelSeed(long seed);
+	public void setLevelSeed(int seed);
 	
 	/*
 	 * Resets generaton count and clears the agent pool, as if we just started anew.
